Add `ClassConversionHandler` abstraction
The `ClassConversionHandler` interface allows to split weaving among
multiple classes.
