+++ log4j-instrument/src/main/java/org/apache/logging/log4j/instrument/LocationClassConverter.java
@@ -14,13 +14,14 @@
  * See the license for the specific language governing permissions and
  * limitations under the license.
  */
-package org.apache.logging.log4j.instrument.location;
+package org.apache.logging.log4j.instrument;
 
 import java.io.IOException;
 import java.io.InputStream;
 import java.io.OutputStream;
 import java.lang.instrument.ClassFileTransformer;
 
+import org.apache.logging.log4j.instrument.log4j2.LoggerConversionHandler;
 import org.objectweb.asm.ClassReader;
 import org.objectweb.asm.ClassWriter;
 
@@ -36,6 +37,7 @@ public class LocationClassConverter implements ClassFileTransformer {
     public void convert(InputStream src, OutputStream dest, LocationCache locationCache) throws IOException {
         final ClassWriter writer = new ClassWriter(ClassWriter.COMPUTE_FRAMES);
         final LocationClassVisitor converter = new LocationClassVisitor(writer, locationCache);
+        converter.addClassConversionHandler(new LoggerConversionHandler());
         new ClassReader(src).accept(converter, ClassReader.EXPAND_FRAMES);
 
         dest.write(writer.toByteArray());
