+++ log4j-maven-plugin/src/main/java/org/apache/logging/log4j/maven/scan/SimpleInclusionScanner.java
@@ -25,7 +25,7 @@ import java.util.Set;
 import java.util.stream.Collectors;
 
 import org.apache.logging.log4j.instrument.Constants;
-import org.apache.logging.log4j.instrument.location.LocationCache;
+import org.apache.logging.log4j.instrument.LocationCache;
 import org.apache.maven.plugin.logging.Log;
 import org.codehaus.plexus.util.DirectoryScanner;
